+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Rename Netdev ip6addr to ipaddr6 for consistency
   - Set `addr-gen-mode=eui64`
 - Set dnsmasq to listen to the Warewulf interface to prevent to binding to localhost:53
+- Added Ipv6 support to `/etc/hosts` on host and nodes.
 
 ### Removed